Issue #7: The main window is supressed
