+++ Clipboard History/Clipboard_HistoryApp.swift
@@ -5,46 +5,6 @@ import Cocoa
 struct Clipboard_HistoryApp: App {
     @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
     var body: some Scene {
-        WindowGroup {
-            ClipboardView(viewModel: ClipboardViewModel())
-                .onAppear {
-                    NSApp.setActivationPolicy(.prohibited)
-                }
-        }
+        WindowGroup {}
     }
 }
-
-
-
-
-
-class AppDelegate: NSObject, NSApplicationDelegate {
-    
-    var statusItem: NSStatusItem!
-    var popover: NSPopover!
-    
-    func applicationDidFinishLaunching(_ aNotification: Notification) {
-        statusItem = NSStatusBar.system.statusItem(withLength: NSStatusItem.squareLength)
-        if let button = statusItem.button {
-            button.image = NSImage(systemSymbolName: "circle", accessibilityDescription: nil)
-            button.action = #selector(togglePopover(_:))
-        }
-        
-        let contentView = ClipboardView(viewModel: ClipboardViewModel())
-        popover = NSPopover()
-        popover.contentSize = NSSize(width: 400, height: 600)
-        popover.behavior = .transient
-        popover.contentViewController = NSHostingController(rootView: contentView)
-    }
-
-    @objc func togglePopover(_ sender: Any?) {
-        if let button = statusItem.button {
-            if popover.isShown {
-                popover.performClose(sender)
-            } else {
-                popover.show(relativeTo: button.bounds, of: button, preferredEdge: .minY)
-            }
-        }
-    }
-}
-
